backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.5,<5.0
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
twilio>=9.2,<10.0
requests>=2.32,<3.0
qrcode>=7.4,<8.0
Pillow>=11.0,<12.0
geopy>=2.4,<3.0
